Followed TechWithTim tutorial
I followed a Tech with Tim tutorial on python sockets. Right now there is a server that listens on a port for connections and messages and handles clients disconnecting. Clients can connect to the server if they have the address and port and can send messages. Client can receive a 'seen' confirmation message from the server.